% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfrgPG.R
\docType{data}
\name{cfrgPG}
\alias{cfrgPG}
\title{Chief Forester's Reference Guide for stocking standards in PG Region}
\format{A data frame with 10763 rows and 17 variables}
\usage{
cfrgPG
}
\description{
A dataset containing tree species suitability ratins for BGC units in Prince George Region.
}
\keyword{datasets}
