% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Clean RESULTS data extracted using extractData function()}
\usage{
cleanData(extractOut)
}
\arguments{
\item{extractOut}{output from extractData().}
}
\description{
This function takes the output from extractData() and prepares it for summaries.
}
\keyword{}
