% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{Extract silviculture data for species and region}
\usage{
extractData(regionCode, speciesList)
}
\arguments{
\item{regionCode}{Three-letter region code (e.g., "ROM").}

\item{speciesList}{List of species codes to be extracted (e.g., c("FD","FDI"))}
}
\description{
This function uses the bcdata package to extract RESULTS entries for
given species and region.
}
\details{
Function can take a VERY long time to run, so recommended to filter by single region at a time
and do not filter for commonly planted species (e.g., Pli).
}
