% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rskFDI.R
\docType{data}
\name{rskFDI}
\alias{rskFDI}
\title{Polygons in Skeena Region planted with Douglas-fir}
\format{A data frame with 602 rows and 66 columns.}
\usage{
rskFDI
}
\description{
A dataset containing silviculture information for polygons planted with Douglas-fir in the Skeena Region.
}
\details{
Dataset was generated using extractData() function.
}
\keyword{datasets}
